Move preferences to their own package
