+++ app/src/main/java/org/greenstand/android/TreeTracker/di/AppModule.kt
@@ -12,36 +12,11 @@ import org.greenstand.android.TreeTracker.background.SyncNotificationManager
 import org.greenstand.android.TreeTracker.managers.LanguageSwitcher
 import org.greenstand.android.TreeTracker.managers.LocationDataCapturer
 import org.greenstand.android.TreeTracker.managers.LocationUpdateManager
-import org.greenstand.android.TreeTracker.managers.Preferences
 import org.greenstand.android.TreeTracker.managers.UserManager
-import org.greenstand.android.TreeTracker.usecases.BundleTreeUploadStrategy
-import org.greenstand.android.TreeTracker.usecases.CreateFakeTreesUseCase
-import org.greenstand.android.TreeTracker.usecases.CreatePlanterCheckInUseCase
-import org.greenstand.android.TreeTracker.usecases.CreatePlanterInfoUseCase
-import org.greenstand.android.TreeTracker.usecases.CreateTreeRequestUseCase
-import org.greenstand.android.TreeTracker.usecases.CreateTreeUseCase
-import org.greenstand.android.TreeTracker.usecases.DeleteOldPlanterImagesUseCase
-import org.greenstand.android.TreeTracker.usecases.ExpireCheckInStatusUseCase
-import org.greenstand.android.TreeTracker.usecases.PlanterCheckInUseCase
-import org.greenstand.android.TreeTracker.usecases.RemoveLocalTreeImagesWithIdsUseCase
-import org.greenstand.android.TreeTracker.usecases.SyncDataUseCase
-import org.greenstand.android.TreeTracker.usecases.TreeUploadStrategy
-import org.greenstand.android.TreeTracker.usecases.UploadImageUseCase
-import org.greenstand.android.TreeTracker.usecases.UploadLocationDataUseCase
-import org.greenstand.android.TreeTracker.usecases.UploadPlanterCheckInUseCase
-import org.greenstand.android.TreeTracker.usecases.UploadPlanterInfoUseCase
-import org.greenstand.android.TreeTracker.usecases.UploadPlanterUseCase
-import org.greenstand.android.TreeTracker.usecases.UploadTreeBundleUseCase
-import org.greenstand.android.TreeTracker.usecases.ValidateCheckInStatusUseCase
+import org.greenstand.android.TreeTracker.preferences.Preferences
+import org.greenstand.android.TreeTracker.usecases.*
 import org.greenstand.android.TreeTracker.utilities.DeviceUtils
-import org.greenstand.android.TreeTracker.viewmodels.DataViewModel
-import org.greenstand.android.TreeTracker.viewmodels.LoginViewModel
-import org.greenstand.android.TreeTracker.viewmodels.MapViewModel
-import org.greenstand.android.TreeTracker.viewmodels.NewTreeViewModel
-import org.greenstand.android.TreeTracker.viewmodels.SignupViewModel
-import org.greenstand.android.TreeTracker.viewmodels.TermsPolicyViewModel
-import org.greenstand.android.TreeTracker.viewmodels.TreeHeightViewModel
-import org.greenstand.android.TreeTracker.viewmodels.TreePreviewViewModel
+import org.greenstand.android.TreeTracker.viewmodels.*
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.viewmodel.dsl.viewModel
 import org.koin.dsl.module
