+++ app/src/main/java/org/greenstand/android/TreeTracker/preferences/PrefKeys.kt
@@ -1,0 +1,10 @@
+package org.greenstand.android.TreeTracker.preferences
+
+object PrefKeys {
+
+    val ROOT = PrefKey("greenstand")
+
+    val SESSION = PrefKey("session")
+
+    val USER_SETTINGS = ROOT + PrefKey("user-settings")
+}
