+++ app/src/main/java/org/greenstand/android/TreeTracker/preferences/UserPrefKey.kt
@@ -1,0 +1,3 @@
+package org.greenstand.android.TreeTracker.preferences
+
+class UserPrefKey(path: String) : PrefKey(path)
